+++ compositor_common/src/scene/validation/constraints/input_count.rs
